move shared api code to one main plugin
