--- client/ayon_harmony/api/plugins/publish/__init__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Publish plugins package.""" 
-
-from .render import CollectRenderBase
-
-__all__ = [
-    "CollectRenderBase",
-]
